scouting and fixing energy shiz :D

+++ src/main.ts
@@ -10,7 +10,9 @@ export function loop(): void {
     if(!Game.creeps[creepName]) {
       console.log(`Deleting memory for dead creep: ${creepName}`)
 
-      deleteCreepMemory(Memory.creeps[creepName].target, Memory.creeps[creepName].workRoom)
+      if (Memory.creeps[creepName].role === 'Hauler' || Memory.creeps[creepName].role === 'Miner'){
+        deleteCreepMemory(Memory.creeps[creepName].target, Memory.creeps[creepName].workRoom)
+      }
 
       delete Memory.creeps[creepName];
     }
